Refactor to extract a split and task mixin
